added extractUser middleware

+++ controllers/blog.js
@@ -1,8 +1,5 @@
 const blogRouter = require("express").Router();
 const Blog = require("../models/blog");
-const User = require("../models/user");
-const jwt = require("jsonwebtoken");
-const config = require("../utils/config");
 
 blogRouter.get("/", async (req, res) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
@@ -11,20 +8,13 @@ blogRouter.get("/", async (req, res) => {
 
 blogRouter.post("/", async (req, res) => {
   const body = req.body;
-  const token = req.token;
-  const decodedToken = jwt.verify(token, config.SECRET);
+  const user = req.user;
 
-  if(!decodedToken.id) {
-    return res.status(401).json({ error: "Token missing or invalid" });
-  }
-
-  if (!body || !body.title || !body.author || !body.url) {
+  if (!(body && body.title && body.author && body.url)) {
     return res.status(400).end();
   }
 
   if (!body.likes) body.likes = 0;
-
-  const user = await User.findById(decodedToken.id);
 
   const blog = new Blog({
     title: body.title,
@@ -58,9 +48,17 @@ blogRouter.get("/:id", async (req, res, next) => {
 blogRouter.put("/:id", async (req, res, next) => {
   try {
     const body = req.body;
+    const userId = req.user;
+    const blogToUpdate = await Blog.findById(req.params.id);
 
     if (!body || !body.title || !body.author || !body.url) {
       return res.status(400).end();
+    }
+
+    if(!(userId && blogToUpdate.user.toString() === userId.toString() )) {
+      return res.status(401).json({
+        error: "Unauthorized blog update request"
+      });
     }
 
     const blog = {
@@ -74,7 +72,6 @@ blogRouter.put("/:id", async (req, res, next) => {
       new: true,
     });
     res.json(updatedBlog);
-    console.log("Updated blog", updatedBlog);
   } catch (err) {
     res.status(400);
     next(err);
@@ -83,6 +80,15 @@ blogRouter.put("/:id", async (req, res, next) => {
 
 blogRouter.delete("/:id", async (req, res, next) => {
   try {
+    const userId = req.user.id;
+    const blog = await Blog.findById(req.params.id);
+    
+    if(!(userId && (blog.user.toString() === userId.toString()))) {
+      return res.status(401).json({
+        error: "Not authorized to delete this blog"
+      });
+    }
+
     await Blog.findByIdAndDelete(req.params.id);
     res.status(204).end();
   } catch (err) {
